VACMS-12787 Move facility events to the appropriate facility module.
