+++ docroot/modules/custom/va_gov_backend/src/EventSubscriber/EntityEventSubscriber.php
@@ -2,22 +2,13 @@
 
 namespace Drupal\va_gov_backend\EventSubscriber;
 
-use Drupal\Component\Render\FormattableMarkup;
 use Drupal\core_event_dispatcher\EntityHookEvents;
 use Drupal\core_event_dispatcher\Event\Entity\EntityPresaveEvent;
 use Drupal\core_event_dispatcher\Event\Entity\EntityTypeBuildEvent;
-use Drupal\core_event_dispatcher\Event\Entity\EntityViewAlterEvent;
-use Drupal\core_event_dispatcher\Event\Form\FormAlterEvent;
 use Drupal\core_event_dispatcher\Event\Form\FormIdAlterEvent;
-use Drupal\core_event_dispatcher\FormHookEvents;
 use Drupal\Core\Config\Entity\ConfigEntityType;
-use Drupal\Core\Entity\ContentEntityForm;
-use Drupal\Core\Entity\EntityFieldManager;
-use Drupal\Core\Entity\EntityFormInterface;
 use Drupal\Core\Entity\EntityTypeManager;
-use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
-use Drupal\Core\StringTranslation\TranslationInterface;
 use Drupal\field_event_dispatcher\Event\Field\WidgetSingleElementFormAlterEvent;
 use Drupal\field_event_dispatcher\FieldHookEvents;
 use Drupal\node\NodeInterface;
@@ -26,7 +17,7 @@ use Drupal\va_gov_user\Service\UserPermsService;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
- * VA.gov VAMC Entity Event Subscriber.
+ * VA.gov Generic Entity Event Subscriber. Do only multi-product stuff here.
  */
 class EntityEventSubscriber implements EventSubscriberInterface {
 
@@ -48,35 +39,33 @@ class EntityEventSubscriber implements EventSubscriberInterface {
   private $entityTypeManager;
 
   /**
-   * The entity field manager.
-   *
-   * @var \Drupal\Core\Entity\EntityFieldManager
-   *  The entity field manager.
-   */
-  private $entityFieldManager;
-
-  /**
    * Constructs the EventSubscriber object.
    *
-   * @param \Drupal\Core\StringTranslation\TranslationInterface $string_translation
-   *   The string translation service.
    * @param \Drupal\va_gov_user\Service\UserPermsService $user_perms_service
    *   The current user perms service.
    * @param \Drupal\Core\Entity\EntityTypeManager $entity_type_manager
-   *   The string entity type service.
-   * @param \Drupal\Core\Entity\EntityFieldManager $entity_field_manager
-   *   The entity field service.
+   *   The entity type manager service.
    */
   public function __construct(
-    TranslationInterface $string_translation,
     UserPermsService $user_perms_service,
     EntityTypeManager $entity_type_manager,
-    EntityFieldManager $entity_field_manager
   ) {
-    $this->stringTranslation = $string_translation;
     $this->userPermsService = $user_perms_service;
     $this->entityTypeManager = $entity_type_manager;
-    $this->entityFieldManager = $entity_field_manager;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function getSubscribedEvents(): array {
+    return [
+      'hook_event_dispatcher.form_node_centralized_content_edit_form.alter' => 'alterCentralizedContentNodeForm',
+      'hook_event_dispatcher.form_node_person_profile_edit_form.alter' => 'alterStaffProfileNodeForm',
+      'hook_event_dispatcher.form_node_person_profile_form.alter' => 'alterStaffProfileNodeForm',
+      EntityHookEvents::ENTITY_PRE_SAVE => 'entityPresave',
+      EntityHookEvents::ENTITY_TYPE_BUILD => 'entityTypeBuild',
+      FieldHookEvents::WIDGET_SINGLE_ELEMENT_FORM_ALTER => 'formWidgetAlter',
+    ];
   }
 
   /**
@@ -103,37 +92,6 @@ class EntityEventSubscriber implements EventSubscriberInterface {
   }
 
   /**
-   * Alteration to entity view pages.
-   *
-   * @param \Drupal\core_event_dispatcher\Event\Entity\EntityViewAlterEvent $event
-   *   The entity view alter service.
-   */
-  public function entityViewAlter(EntityViewAlterEvent $event):void {
-    $this->showUnspecifiedWhenSystemEhrNumberEmpty($event);
-  }
-
-  /**
-   * Alterations to Vet center node forms.
-   *
-   * @param \Drupal\core_event_dispatcher\Event\Form\FormIdAlterEvent $event
-   *   The event.
-   */
-  public function alterVetCenterNodeForm(FormIdAlterEvent $event): void {
-    $this->buildHealthServicesDescriptionArrayAddToSettings($event);
-  }
-
-  /**
-   * Alterations to VAMC system health service node forms.
-   *
-   * @param \Drupal\core_event_dispatcher\Event\Form\FormIdAlterEvent $event
-   *   The event.
-   */
-  public function alterRegionalHealthServiceNodeForm(FormIdAlterEvent $event): void {
-    $this->buildRegionalHealthServiceFormIntro($event);
-    $this->buildHealthServicesDescriptionArrayAddToSettings($event);
-  }
-
-  /**
    * Overrides Block Content Type access handler.
    *
    * @param \Drupal\core_event_dispatcher\Event\Entity\EntityTypeBuildEvent $event
@@ -149,177 +107,6 @@ class EntityEventSubscriber implements EventSubscriberInterface {
     $entityTypes = &$event->getEntityTypes();
     if (!empty($entityTypes['block_content_type']) && $entityTypes['block_content_type'] instanceof ConfigEntityType) {
       $entityTypes['block_content_type']->setAccessClass(BlockContentTypeAccessControlHandler::class);
-    }
-  }
-
-  /**
-   * Builds h2 of VAMC System Health Service page type name and adds help text.
-   *
-   * @param \Drupal\core_event_dispatcher\Event\Form\FormIdAlterEvent $event
-   *   The event.
-   */
-  public function buildRegionalHealthServiceFormIntro(FormIdAlterEvent $event): void {
-    $form = &$event->getForm();
-    $formatted_markup = new FormattableMarkup('<div class="services-intro-wrap"><h2>VAMC System Health Service</h2>
-    <p>Add services that Veterans can receive at one or more facilities in your health system.
-    Some content won’t be editable because it comes from other sources. For full guidance,
-    see <a target="_blank" href="@help_link">How to edit a VAMC System Health Service (opens in a new tab)</a>.</p></div>', [
-      '@help_link' => 'https://prod.cms.va.gov/help/vamc/how-to-add-a-vamc-system-health-service',
-    ]);
-    $form['field_service_name_and_descripti']['#prefix'] = $this->t('@markup', ['@markup' => $formatted_markup]);
-  }
-
-  /**
-   * Builds an array of descriptions from health services available on form.
-   *
-   * Adds the descriptions array built by this method to drupalSettings.
-   *
-   * @param \Drupal\core_event_dispatcher\Event\Form\FormIdAlterEvent $event
-   *   The event.
-   */
-  public function buildHealthServicesDescriptionArrayAddToSettings(FormIdAlterEvent $event): void {
-    $form = &$event->getForm();
-    $form_state = $event->getFormState();
-    $entity_type = 'taxonomy_term';
-    $bundle = 'health_care_service_taxonomy';
-    $fields = $this->getProductTypeTermFields($form, $form_state);
-    $service_terms = $this->entityTypeManager
-      ->getListBuilder($entity_type)
-      ->getStorage()
-      ->loadByProperties([
-        'vid' => $bundle,
-      ]);
-    // Use this to grab values in the term parent vocab.
-    $vocabulary_definition = $this->entityFieldManager->getFieldDefinitions($entity_type, $bundle);
-    $descriptions = [];
-    foreach ($service_terms as $service_term) {
-      $description = $service_term->get($fields['description'])->value ?? '';
-      // Lovell VAMC Facilities may include a TRICARE description field.
-      $tricare_description = $service_term->get('field_tricare_description')->value ?? '';
-      /** @var \Drupal\taxonomy\Entity\Term $service_term */
-      $descriptions[$service_term->id()] = [
-        'type' => $service_term->get($fields['type'])->getSetting('allowed_values')[$service_term->get($fields['type'])->getString()] ?? NULL,
-        'name' => $service_term->get($fields['name'])->getString(),
-        'conditions' => $service_term->get($fields['conditions'])->getString(),
-        'description' => trim(strip_tags($description)),
-        'tricare_description' => trim(strip_tags($tricare_description)),
-        'vc_vocabulary_service_description_label' => $vocabulary_definition['field_vet_center_service_descrip']->getLabel(),
-        'vc_vocabulary_description_help_text' => $vocabulary_definition['field_vet_center_service_descrip']->getDescription(),
-      ];
-    }
-    $form['#attached']['drupalSettings']['availableHealthServices'] = $descriptions;
-    $form['#attached']['library'][] = 'va_gov_backend/display_service_descriptions';
-  }
-
-  /**
-   * Form alter Event call.
-   *
-   * @param \Drupal\core_event_dispatcher\Event\Form\FormAlterEvent $event
-   *   The event.
-   */
-  public function formAlter(FormAlterEvent $event): void {
-    $form = &$event->getForm();
-    $form_state = $event->getFormState();
-    $this->lockTitleEditing($form, $form_state);
-    $this->lockApiIdEditing($form, $form_state);
-    $this->removeDeleteButton($form);
-  }
-
-  /**
-   * Builds an array of term fields predicated by product type.
-   *
-   * @param array $form
-   *   The form array.
-   * @param \Drupal\Core\Form\FormStateInterface $form_state
-   *   The form state.
-   */
-  public function getProductTypeTermFields(array &$form, FormStateInterface $form_state) {
-    $fields = [];
-    if ($form_state->getFormObject() instanceof EntityFormInterface) {
-      $bundle = $form_state->getFormObject()->getEntity()->bundle();
-      // Make the bundle available to displayServiceDescriptions.js.
-      $form['#attached']['drupalSettings']['currentNodeBundle'] = $bundle;
-      $fields = [
-        'type' => $bundle === 'vet_center' ? 'field_vet_center_type_of_care' : 'field_service_type_of_care',
-        'name' => $bundle === 'vet_center' ? 'field_vet_center_friendly_name' : 'field_also_known_as',
-        'conditions' => $bundle === 'vet_center' ? 'field_vet_center_com_conditions' : 'field_commonly_treated_condition',
-        'description' => $bundle === 'vet_center' ? 'field_vet_center_service_descrip' : 'description',
-      ];
-    }
-    return $fields;
-  }
-
-  /**
-   * Shows the text "Unspecified" when phone number is blank.
-   *
-   * @param \Drupal\core_event_dispatcher\Event\Entity\EntityViewAlterEvent $event
-   *   The entity view alter service.
-   */
-  public function showUnspecifiedWhenSystemEhrNumberEmpty(EntityViewAlterEvent $event):void {
-    if ($event->getDisplay()->getTargetBundle() === 'health_care_region_page') {
-      $build = &$event->getBuild();
-      if (empty($build['field_va_health_connect_phone']['#title'])) {
-        $undefined_number_text = '
-          <div class="field field--name-field-va-health-connect-phone field--type-list-string field--label-above">
-              <div class="field__label">VA Health Connect phone number</div>
-              <div class="field__item">Undefined</div>
-          </div>';
-
-        $formatted_markup = new FormattableMarkup($undefined_number_text, []);
-        $build['field_va_health_connect_phone']['#prefix'] = $formatted_markup;
-      }
-    }
-
-  }
-
-  /**
-   * Locks down standardized form titles for non-admins.
-   *
-   * @param array $form
-   *   The form.
-   * @param \Drupal\Core\Form\FormStateInterface $form_state
-   *   The form state.
-   */
-  public function lockTitleEditing(array &$form, FormStateInterface $form_state) {
-    $form_object = $form_state->getFormObject();
-    $bundle = NULL;
-    if ($form_object instanceof ContentEntityForm) {
-      $bundle = $form_object->getEntity()->bundle();
-    }
-    $bundles_with_standardized_titles = [
-      'event_listing',
-      'health_services_listing',
-      'leadership_listing',
-      'locations_listing',
-      'office',
-      'press_releases_listing',
-      'publication_listing',
-      'story_listing',
-    ];
-    if (!$this->userPermsService->hasAdminRole() && in_array($bundle, $bundles_with_standardized_titles)) {
-      $form['title']['#disabled'] = TRUE;
-    }
-  }
-
-  /**
-   * Locks down API Id for non-admins.
-   *
-   * @param array $form
-   *   The form.
-   * @param \Drupal\Core\Form\FormStateInterface $form_state
-   *   The form state.
-   */
-  public function lockApiIdEditing(array &$form, FormStateInterface $form_state) {
-    $form_object = $form_state->getFormObject();
-    $bundle = NULL;
-    if ($form_object instanceof ContentEntityForm) {
-      $bundle = $form_object->getEntity()->bundle();
-    }
-    if (!$this->userPermsService->hasAdminRole(TRUE) && $bundle === 'health_care_service_taxonomy') {
-      $form['field_health_service_api_id']['#disabled'] = TRUE;
-    }
-    if (!$this->userPermsService->hasAdminRole(TRUE) && $bundle === 'facility_supplemental_status') {
-      $form['field_status_id']['#disabled'] = TRUE;
     }
   }
 
@@ -483,7 +270,6 @@ class EntityEventSubscriber implements EventSubscriberInterface {
       else {
         unset($form['field_product']);
       }
-
       // For non-admin replace applied to field with static content
       // generated from field value and disable editing.
       if (!empty($form['field_applied_to']['widget'][0]['value']['#default_value'])) {
@@ -523,41 +309,6 @@ class EntityEventSubscriber implements EventSubscriberInterface {
     if (!empty($form['#paragraph_type']) && $form['#paragraph_type'] === 'button') {
       unset($form['top']['actions']['actions']['collapse_button']);
     }
-  }
-
-  /**
-   * Remove delete button from edit forms.
-   *
-   * @param array $form
-   *   The form.
-   */
-  public function removeDeleteButton(array &$form) {
-    // Users with any role, other than administrator role,
-    // shouldn't be able to delete anything.
-    if (!$this->userPermsService->hasAdminRole(TRUE) && isset($form['actions']['delete'])) {
-      // Remove the delete button.
-      $form['actions']['delete']['#access'] = FALSE;
-    }
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function getSubscribedEvents(): array {
-    return [
-      EntityHookEvents::ENTITY_PRE_SAVE => 'entityPresave',
-      EntityHookEvents::ENTITY_TYPE_BUILD => 'entityTypeBuild',
-      FieldHookEvents::WIDGET_SINGLE_ELEMENT_FORM_ALTER => 'formWidgetAlter',
-      EntityHookEvents::ENTITY_VIEW_ALTER => 'entityViewAlter',
-      FormHookEvents::FORM_ALTER => 'formAlter',
-      'hook_event_dispatcher.form_node_person_profile_form.alter' => 'alterStaffProfileNodeForm',
-      'hook_event_dispatcher.form_node_person_profile_edit_form.alter' => 'alterStaffProfileNodeForm',
-      'hook_event_dispatcher.form_node_centralized_content_edit_form.alter' => 'alterCentralizedContentNodeForm',
-      'hook_event_dispatcher.form_node_vet_center_form.alter' => 'alterVetCenterNodeForm',
-      'hook_event_dispatcher.form_node_vet_center_edit_form.alter' => 'alterVetCenterNodeForm',
-      'hook_event_dispatcher.form_node_regional_health_care_service_des_form.alter' => 'alterRegionalHealthServiceNodeForm',
-      'hook_event_dispatcher.form_node_regional_health_care_service_des_edit_form.alter' => 'alterRegionalHealthServiceNodeForm',
-    ];
   }
 
 }
