+++ docroot/modules/custom/va_gov_backend/va_gov_backend.libraries.yml
@@ -23,15 +23,6 @@ contact_paragraphs_interactions:
   version: 1.x
   js:
     js/contact_paragraphs_interactions.js: {}
-display_service_descriptions:
-  js:
-    js/displayServiceDescriptions.js: {}
-  css:
-    component:
-      css/css_tooltip.css: {}
-      css/health_service.css: {}
-  dependencies:
-    - va_gov_lovell/winnow_service_names_tricare
 field_group_tooltip:
   version: 1.x
   js:
