+++ docroot/modules/custom/va_gov_backend/va_gov_backend.services.yml
@@ -1,7 +1,7 @@
 services:
   va_gov_backend.entity_event_subscriber:
     class: Drupal\va_gov_backend\EventSubscriber\EntityEventSubscriber
-    arguments: ['@string_translation', '@va_gov_user.user_perms', '@entity_type.manager', '@entity_field.manager']
+    arguments: ['@va_gov_user.user_perms', '@entity_type.manager']
     tags:
       - { name: event_subscriber }
   va_gov_backend.table_audit_views_event_subscriber:
