+++ docroot/modules/custom/va_gov_facilities/js/displayServiceDescriptions.js
