+++ docroot/modules/custom/va_gov_facilities/va_gov_facilities.libraries.yml
@@ -1,3 +1,12 @@
+display_service_descriptions:
+  js:
+    js/displayServiceDescriptions.js: {}
+  css:
+    component:
+      css/css_tooltip.css: {}
+      css/health_service.css: {}
+  dependencies:
+    - va_gov_lovell/winnow_service_names_tricare
 facility_status_details:
   version: 1.x
   js:
