+++ docroot/modules/custom/va_gov_facilities/va_gov_facilities.services.yml
@@ -1,6 +1,12 @@
 services:
   va_gov_facilities.entity_event_subscriber:
     class: Drupal\va_gov_facilities\EventSubscriber\FacilitiesSubscriber
-    arguments: ['@current_user', '@entity_type.manager', '@messenger']
+    arguments:
+      - '@current_user'
+      - '@entity_field.manager'
+      - '@entity_type.manager'
+      - '@messenger'
+      - '@string_translation'
+      - '@va_gov_user.user_perms'
     tags:
       - { name: event_subscriber }
